Use express middleware in openstad auth - use express middleware with when: afterRequired property (after all middleware shipped with apostrophe-express but before all middleware passed to it as options) - move logic to lib folder
